Enum nft_marketplace::MarketAction

pub enum MarketAction {
    AddNftContract(ActorId),
    AddFTContract(ActorId),
    AddMarketData {
        nft_contract_id: ActorId,
        ft_contract_id: Option<ActorId>,
        token_id: U256,
        price: Option<u128>,
    },
    BuyItem {
        nft_contract_id: ActorId,
        token_id: U256,
    },
    CreateAuction {
        nft_contract_id: ActorId,
        ft_contract_id: Option<ActorId>,
        token_id: U256,
        min_price: u128,
        bid_period: u64,
        duration: u64,
    },
    AddBid {
        nft_contract_id: ActorId,
        token_id: U256,
        price: u128,
    },
    SettleAuction {
        nft_contract_id: ActorId,
        token_id: U256,
    },
    AddOffer {
        nft_contract_id: ActorId,
        ft_contract_id: Option<ActorId>,
        token_id: U256,
        price: u128,
    },
    Withdraw {
        nft_contract_id: ActorId,
        token_id: U256,
        hash: H256,
    },
    AcceptOffer {
        nft_contract_id: ActorId,
        token_id: U256,
        offer_hash: H256,
    },
    Item {
        nft_contract_id: ActorId,
        token_id: U256,
    },
}

Variants

AddNftContract(ActorId)

Adds NFT contract addresses that can be listed on marketplace.

Requirements:

Only admin can add approved NFT accounts.

Arguments:

nft_contract_id: the NFT contract address

AddFTContract(ActorId)

Adds the contract addresses of fungible tokens with which users can pay for NFTs.

Requirements:

Only admin can add approved fungible-token accounts.

Arguments:

ft_contract_id: the FT contract address

AddMarketData

Fields

nft_contract_id: ActorId

ft_contract_id: Option<ActorId>

token_id: U256

price: Option<u128>

Adds data on market item. If the item of that NFT does not exist on the marketplace then it will be listed. If the item exists then that action is used to change the price or suspend the sale.

Requirements

msg::source() must be the NFT owner

nft_contract_id must be in the list of approved_nft_contracts

if item already exists, then it cannot be changed if there is an active auction

Arguments:

nft_contract_id: the NFT contract address

token_id: the NFT id

price: the NFT price (if it is None then the item is not on the sale)

On success replies MarketEvent::MarketDataAdded.

BuyItem

Fields

nft_contract_id: ActorId

token_id: U256

Sells the NFT.

Requirements:

The NFT item must exists and be on sale.

If the NFT is sold for a native Gear value, then a buyer must attach value equals to the price.

If the NFT is sold for fungible tokens then a buyer must have enough tokens in the fungible token contract.

There must be no an opened auction on the item.

Arguments:

nft_contract_id: NFT contract address

token_id: the token ID

On success replies MarketEvent::ItemSold.

CreateAuction

Fields

nft_contract_id: ActorId

ft_contract_id: Option<ActorId>

token_id: U256

min_price: u128

bid_period: u64

duration: u64

Creates an auction for selected item. If the NFT item doesn’t exist on the marketplace then it will be listed

Requirements:

Only the item owner can start auction.

nft_contract_id must be in the list of approved_nft_contracts

There must be no active auction.

Arguments:

nft_contract_id: the NFT contract address

ft_contract_id: the fungible token contract address that can be used for trading

token_id: the NFT id

min_price: the starting price

bid_period: the time interval. If the auction ends before exec::blocktimestamp() + bid_period then the auction end time is delayed for bid_period.

On success replies MarketEvent::AuctionCreated.

AddBid

Fields

nft_contract_id: ActorId

token_id: U256

price: u128

Adds a bid to an ongoing auction.

Requirements:

The item must extsts.

The auction must exists on the item.

If the NFT is sold for a native Gear value, then a buyer must attach value equals to the price indicated in the arguments.

If the NFT is sold for fungible tokens then a buyer must have enough tokens in the fungible token contract.

price must be greater then the current offered price for that item.

Arguments

nft_contract_id: the NFT contract address.

token_id: the NFT id.

price: the offered price.

On success replies MarketEvent::BidAdded.

SettleAuction

Fields

nft_contract_id: ActorId

token_id: U256

Settles the auction.

Requirements:

The auction must be over.

Arguments:

nft_contract_id: the NFT contract address

token_id: the NFT id

On successful auction replies MarketEvent::AuctionSettled. If no bids were made replies MarketEvent::AuctionCancelled.

AddOffer

Fields

nft_contract_id: ActorId

ft_contract_id: Option<ActorId>

token_id: U256

price: u128

Adds a price offer to the item.

Requirements:

NFT item must exists and be listed on the marketplace.

There must be no an ongoing auction on the item.

If a user makes an offer in native Gear value, then he must attach value equals to the price indicated in the arguments.

If a user makes an offer in fungible tokens then he must have enough tokens in the fungible token contract.

The price can not be equal to 0.

There must be no identical offers on the item.

Arguments:

nft_contract_id: the NFT contract address

ft_contract_id: the FT contract address

token_id: the NFT id

price: the offer price

On success replies MarketEvent::OfferAdded.

Withdraw

Fields

nft_contract_id: ActorId

token_id: U256

hash: H256

Withdraws tokens.

Requirements:

NFT item must exists and be listed on the marketplace.

Only the offer creator can withdraw his tokens.

The offer with indicated hash must exist.

Arguments:

nft_contract_id: the NFT contract address

token_id: the NFT id

offer_hash: the offer hash that includes the offer price and the address of fungible token contract.

On success replies MarketEvent::TokensWithdrawn.

AcceptOffer

Fields

nft_contract_id: ActorId

token_id: U256

offer_hash: H256

Accepts an offer.

Requirements:

NFT item must exists and be listed on the marketplace.

Only owner can accept offer.

There must be no ongoing auction.

The offer with indicated hash must exist.

Arguments:

nft_contract_id: the NFT contract address

token_id: the NFT id

offer_hash: the offer hash that includes the offer price and the address of fungible token contract.

On success replies MarketEvent::OfferAccepted.

Item

Fields

nft_contract_id: ActorId

token_id: U256